Move quality data to accordeon
